Module pool_xyk::pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

Structs

Pallet

The pallet implementing the on-chain logic.

Enums

Call

Contains one variant per dispatchable that can be called by an extrinsic.

Error

Custom dispatch errors of this pallet.

Event

The event emitted by this pallet.

Traits

Config

Configuration trait of this pallet.

Type Definitions

AccountPools

Set of pools in which accounts have some share. Liquidity provider account => Target Asset of pair (assuming base asset is XOR)

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.

PoolProviders

Liquidity providers of particular pool. Pool account => Liquidity provider account => Pool token balance

Properties

Properties of particular pool. Base Asset => Target Asset => (Reserves Account Id, Fees Account Id)

Reserves

Updated after last liquidity change operation. [Base Asset Id (XOR) -> Target Asset Id => (Base Balance, Target Balance)]. This storage records is not used as source of information, but used as quick cache for information that comes from balances for assets from technical accounts. For example, communication with technical accounts and their storage is not needed, and this pair to balance cache can be used quickly.

TotalIssuances

Total issuance of particular pool. Pool account => Total issuance