Trait pool_xyk::pallet::Config

pub trait Config: Config + Config + Config + Config + Config + Config {

    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type PairSwapAction: SwapAction<AccountIdOf<Self>, <Self as Config>::TechAccountId, AssetIdOf<Self>, Self> + Parameter;
    type DepositLiquidityAction: SwapAction<AccountIdOf<Self>, <Self as Config>::TechAccountId, AssetIdOf<Self>, Self> + Parameter;
    type WithdrawLiquidityAction: SwapAction<AccountIdOf<Self>, <Self as Config>::TechAccountId, AssetIdOf<Self>, Self> + Parameter;
    type PolySwapAction: SwapAction<AccountIdOf<Self>, <Self as Config>::TechAccountId, AssetIdOf<Self>, Self> + Parameter + Into<<Self as Config>::SwapAction> + From<PolySwapAction<DexIdOf<Self>, AssetIdOf<Self>, <Self as Config>::AccountId, <Self as Config>::TechAccountId>>;
    type EnsureDEXManager: EnsureDEXManager<Self::DEXId, Self::AccountId, DispatchError>;
    type TradingPairSourceManager: TradingPairSourceManager<Self::DEXId, AssetIdOf<Self>>;
    type DexInfoProvider: DexInfoProvider<Self::DEXId, DEXInfo<AssetIdOf<Self>>>;
    type EnabledSourcesManager: EnabledSourcesManager<Self::DEXId, AssetIdOf<Self>>;
    type EnsureTradingPairExists: EnsureTradingPairExists<Self::DEXId, AssetIdOf<Self>, DispatchError>;
    type XSTMarketInfo: GetMarketInfo<AssetIdOf<Self>>;
    type GetFee: Get<Fixed>;
    type OnPoolCreated: OnPoolCreated<AccountId = AccountIdOf<Self>, DEXId = <Self as Config>::DEXId>;
    type OnPoolReservesChanged: OnPoolReservesChanged<AssetIdOf<Self>>;
    type GetTradingPairRestrictedFlag: GetByKey<TradingPair<AssetIdOf<Self>>, bool>;
    type GetChameleonPoolBaseAssetId: GetByKey<AssetIdOf<Self>, Option<AssetIdOf<Self>>>;
    type GetChameleonPool: GetByKey<TradingPair<AssetIdOf<Self>>, bool>;
    type AssetInfoProvider: AssetInfoProvider<AssetIdOf<Self>, Self::AccountId, AssetSymbol, AssetName, BalancePrecision, ContentSource, Description>;
    type IrreducibleReserve: Get<Percent>;
    type WeightInfo: WeightInfo;

    const MIN_XOR: Balance;

}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

Because this pallet emits events, it depends on the runtime’s definition of an event.

type PairSwapAction: SwapAction<AccountIdOf<Self>, <Self as Config>::TechAccountId, AssetIdOf<Self>, Self> + Parameter

type DepositLiquidityAction: SwapAction<AccountIdOf<Self>, <Self as Config>::TechAccountId, AssetIdOf<Self>, Self> + Parameter

type WithdrawLiquidityAction: SwapAction<AccountIdOf<Self>, <Self as Config>::TechAccountId, AssetIdOf<Self>, Self> + Parameter

type PolySwapAction: SwapAction<AccountIdOf<Self>, <Self as Config>::TechAccountId, AssetIdOf<Self>, Self> + Parameter + Into<<Self as Config>::SwapAction> + From<PolySwapAction<DexIdOf<Self>, AssetIdOf<Self>, <Self as Config>::AccountId, <Self as Config>::TechAccountId>>

type EnsureDEXManager: EnsureDEXManager<Self::DEXId, Self::AccountId, DispatchError>

type TradingPairSourceManager: TradingPairSourceManager<Self::DEXId, AssetIdOf<Self>>

type DexInfoProvider: DexInfoProvider<Self::DEXId, DEXInfo<AssetIdOf<Self>>>

type EnabledSourcesManager: EnabledSourcesManager<Self::DEXId, AssetIdOf<Self>>

type EnsureTradingPairExists: EnsureTradingPairExists<Self::DEXId, AssetIdOf<Self>, DispatchError>

type XSTMarketInfo: GetMarketInfo<AssetIdOf<Self>>

type GetFee: Get<Fixed>

type OnPoolCreated: OnPoolCreated<AccountId = AccountIdOf<Self>, DEXId = <Self as Config>::DEXId>

type OnPoolReservesChanged: OnPoolReservesChanged<AssetIdOf<Self>>

type GetTradingPairRestrictedFlag: GetByKey<TradingPair<AssetIdOf<Self>>, bool>

type GetChameleonPoolBaseAssetId: GetByKey<AssetIdOf<Self>, Option<AssetIdOf<Self>>>

type GetChameleonPool: GetByKey<TradingPair<AssetIdOf<Self>>, bool>

type AssetInfoProvider: AssetInfoProvider<AssetIdOf<Self>, Self::AccountId, AssetSymbol, AssetName, BalancePrecision, ContentSource, Description>

To retrieve asset info

type IrreducibleReserve: Get<Percent>

Percent of reserve which is not involved in swap

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.

Required Associated Constants

const MIN_XOR: Balance

The minimum amount of XOR to deposit as liquidity